requirements.txt
git+https://github.com/opentensor/bittensor.git@revolution
torch
pandas
mathgenerator
plotly
nbformat
tqdm
openai
latex2sympy2
antlr4-python3-runtime-4.11.0
wolframalpha